+++ src/Avatar/index.tsx
@@ -1,20 +1,25 @@
 import { Avatar as AntdAvatar, AvatarProps as AntdAvatarProps } from 'antd';
 import React from 'react';
 
-import { MLL_THEME } from '../theme';
-
 import { stringToHslaColor } from './utils';
 
 export const Avatar: typeof AntdAvatar = AntdAvatar;
 export type AvatarProps = AntdAvatarProps;
 
-export function UserAvatar(props: AvatarProps & { username: string }) {
-  const color = stringToHslaColor(props.username);
+export type UserAvatarProps = AvatarProps & { username: string };
+
+export function UserAvatar(props: UserAvatarProps) {
+  const { username, ...rest } = props;
+
   return (
     <Avatar
-      style={{ backgroundColor: color, color: MLL_THEME.backgroundColor }}
+      style={{
+        backgroundColor: stringToHslaColor(username),
+        color: 'white',
+      }}
+      {...rest}
     >
-      {props.username}
+      {username}
     </Avatar>
   );
 }