+++ src/Image/index.tsx
@@ -1,0 +1,5 @@
+import { Image as AntdImage } from 'antd';
+
+export { ImageProps } from 'antd';
+
+export const Image: typeof AntdImage = AntdImage;