+++ src/Plate/index.stories.tsx
@@ -11,13 +11,13 @@ export const Default: Story<Partial<PlateProps>> = (args) => (
   <Plate
     data={[
       {
+        coordinates: { row: 'A', column: 7 },
+        content: <i>It renders any ReactNode</i>,
+      },
+      {
         coordinates: { row: 'A', column: 8 },
         content: 'Test',
         selected: true,
-      },
-      {
-        coordinates: { row: 'A', column: 7 },
-        content: <i>It renders any ReactNode</i>,
       },
       {
         coordinates: { row: 'B', column: 3 },