+++ src/Radio/index.stories.tsx
@@ -1,30 +1,37 @@
 import { Story } from '@storybook/react';
-import { Input } from 'antd';
 import React from 'react';
 
-import { Radio } from './index';
+import { Radio, RadioGroupProps } from './index';
 
 export default {
   title: 'Radio',
-  component: Input,
-};
-
-const options = [
-  { label: 'Apple', value: 'Apple' },
-  { label: 'Pear', value: 'Pear' },
-  { label: 'Orange', value: 'Orange' },
-];
-
-export const Default: Story = (args) => (
-  <Radio.Group options={options} {...args} />
-);
-
-Default.argTypes = {
-  optionType: {
-    control: {
-      type: 'inline-radio',
-      options: ['default', 'button'],
-      default: 'button',
+  component: Radio.Group,
+  argTypes: {
+    onChange: { action: 'onChange' },
+    optionType: {
+      control: {
+        type: 'inline-radio',
+        options: ['default', 'button'],
+        default: 'button',
+      },
+    },
+    buttonStyle: {
+      control: {
+        type: 'inline-radio',
+        options: ['solid', 'outline'],
+        default: 'outline',
+      },
     },
   },
 };
+
+export const Default: Story<RadioGroupProps> = (args) => (
+  <Radio.Group
+    options={[
+      { label: 'Apple', value: 'Apple' },
+      { label: 'Pear', value: 'Pear' },
+      { label: 'Orange', value: 'Orange' },
+    ]}
+    {...args}
+  />
+);